+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.1:
+* Set-NanoLeaf:  Adding -EffectOption parameter help (#9).  Adding examples.
+* Set-NanoLeaf:  Fixing Brightness (#7).  Adding Tab Completion (#8).  Adding Examples
+---
+
 ## 0.2:
 Adding: Disconnect-HueBridge, Disconnect-NanoLeaf, Disconnect-Twinkly (#2)
 Adding Pixoo commands: Connect/Disconnect/Get/Set-Pixoo (#4)